Function infrablockspace_parachain::service::start_shell_node_impl

async fn start_shell_node_impl<RuntimeApi, RB, BIQ, BIC>(
    parachain_config: Configuration,
    infrablockspace_config: Configuration,
    collator_options: CollatorOptions,
    para_id: ParaId,
    rpc_ext_builder: RB,
    build_import_queue: BIQ,
    build_consensus: BIC,
    hwbench: Option<HwBench>
) -> Result<(TaskManager, Arc<TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>>)>where
    RuntimeApi: ConstructRuntimeApi<Block, TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>> + Send + Sync + 'static,
    RuntimeApi::RuntimeApi: TaggedTransactionQueue<Block> + Metadata<Block> + SessionKeys<Block> + ApiExt<Block, StateBackend = StateBackendFor<TFullBackend<Block>, Block>> + OffchainWorkerApi<Block> + BlockBuilder<Block> + CollectCollationInfo<Block>,
    StateBackendFor<TFullBackend<Block>, Block>: StateBackend<BlakeTwo256>,
    RB: Fn(Arc<TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>>) -> Result<RpcModule<()>, Error> + 'static,
    BIQ: FnOnce(Arc<TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>>, TParachainBlockImport<Block, Arc<TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>>, TFullBackend<Block>>, &Configuration, Option<TelemetryHandle>, &TaskManager) -> Result<DefaultImportQueue<Block, TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>>, Error>,
    BIC: FnOnce(Arc<TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>>, TParachainBlockImport<Block, Arc<TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>>, TFullBackend<Block>>, Option<&Registry>, Option<TelemetryHandle>, &TaskManager, Arc<dyn RelayChainInterface>, Arc<FullPool<Block, TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>>>, Arc<SyncingService<Block>>, SyncCryptoStorePtr, bool) -> Result<Box<dyn ParachainConsensus<Block>>, Error>,

Start a shell node with the given parachain Configuration and relay chain Configuration.

This is the actual implementation that is abstract over the executor and the runtime api for shell nodes.